requirements.txt
open3d-python
scipy<=1.1.0
sklearn
h5py
numpy
matplotlib==2.0.0
NearPy
Pillow>=6.2.0
python-dateutil
PyYAML
scikit-image==0.11.3
scikit-learn
six
trimesh
pyhull
cvxopt
pandas
visualization
pyquaternion
numba
autolab-core==0.0.4
autolab-perception
torch>=0.4.1
tensorboardX
mayavi
PyQt4
coloredlogs
